Fix 2fa  number generation bug
Base32 padding are present in apache Base32, but cannot be parsed by
2fa lib properly, so sometimes one odd byte appears during decoding
base32 secret, which has padding. To prevent that, use base32 secret
 without possible padding (in other words -> base32 word should
be of size which is multiple of 8).

Fix validation algorithm, force windowMillis to be 0 in
TimeBasedOneTimePasswordUtil.validateCurrentNumber.
Reason: for example - current time is 970, you generate code, send
to server which generate code at time 980, but you set windowMillis
 - 100 and time_step_millis - 300. So your code will be generated for
time: from 980 - 100 to 980 + 100 with step 300.
In this example code will be generated only for time 880, but expected
at least 880 and 1080.
When you generate code on time 880 but user generate time at time 970,
you has probability to generate wrong code, which will never match
user's code
This is an issue, which cause 2fa code validation problem