Fix startShuffler endpoint for a vault wallet recipient
It is required to specify accountId (in addition to 'passphrase') for every place, where access to the vault wallet is required, but it was an error for HttpParameterParserUtil.getPublicKey
method call, when GetAccountId and ShufflingProcess endpoints, which support vaults, but passed 0 account id to the HttpParameterParserUtil.getPublicKey method (indirectly), which caused a weird error message: 'At least one of :[recipientSecretPhrase, recipientPublicKey, recipientPassphrase] must be specified'. It is a second bug, which is fixed for case, when account id is not specified by adding the following error message: 'account' is not specified.